Updated pylint scores

+++ server.py
@@ -2,16 +2,17 @@
 Implements asynchronous Server to handle multiple clients using event loops
 '''
 
+
 import os
 import asyncio
-from async_server_client import Server
+from test_server import Server
 
 # pathname for unix socket
-UNIX_PATHNAME = ""
+UNIX_PATHNAME = "/home/niranjan/server.socket"
 
 
 # if socket file exists delete it
-if os.path.exists(UNIX_PATHNAME):
+if UNIX_PATHNAME and os.path.exists(UNIX_PATHNAME):
     print("File already exists. Removing it now")
     os.remove(UNIX_PATHNAME)
     print("File removed")
@@ -25,7 +26,7 @@ def handle_data(data, addr, client_sock=None, transport=None):
     client_data = str(addr) + " " + data.decode()
     print("Sending to ", addr, " data ", client_data)
     if client_sock:
-        loop.create_task(s.send(loop, client_sock, client_data))
+        loop.create_task(server.send(loop, client_sock, client_data))
     elif transport:
         transport.sendto(client_data.encode(), addr)
 
@@ -35,14 +36,16 @@ def run():
     Function to start the async server
     '''
     while True:
-        client, addr = loop.run_until_complete(s.accept(loop, handle_data))
+        client, addr = loop.run_until_complete(
+            server.accept(loop, handle_data))
         # print(client, addr)
         if client:
-            loop.create_task(s.recv(loop, client, addr, 4096, handle_data))
+            loop.create_task(server.recv(
+                loop, client, addr, handle_data))
 
 
 if __name__ == "__main__":
-    # s = Server(host="127.0.0.1", port=9000)
-    s = Server(host="127.0.0.1", port=9000)
+    # pylint: disable=invalid-name
+    server = Server(host="127.0.0.1", port=9000, pathname=UNIX_PATHNAME)
     loop = asyncio.get_event_loop()
     run()
